YoloV8_Pose_Fine_Tuning/requirements.txt
ultralytics>=8.0.0
PyYAML
requests
tqdm
numpy
opencv-python
torch>=2.0.0
torchvision>=0.15.0
